user-update-consumer/requirements.txt
pika==1.3.2
mysql-connector-python==8.0.33
defusedxml==0.7.1